device_sdk/python/requirements.txt
# Device SDK (Python) — Dependencies
requests>=2.32.0
paho-mqtt>=2.1.0
